Modify implementation of Upload-Defer-Length
The creation-defer-length extension of the tus resumable upload protocol
version 1.0.0 allows creating upload resources without Upload-Length with
a POST request. The specification does not explicitly enforce the very first
PATCH request to contain an Upload-Length header so resources without
Upload-Length can be filled with data with increasing Upload-Offset.

This logically implies that:
- Upload-Length cannot be lower than current Upload-Offset
- Content-Length + Upload-Offset must not exceed Tus-Max-Size